centralize + refactor types
