+++ apps/web/src/components/task-list-sidebar.tsx
@@ -6,7 +6,8 @@ import {
   PanelRightOpen,
 } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { useTasks, ThreadSummary } from "@/providers/Task";
+import { useTasks } from "@/providers/Task";
+import { ThreadSummary } from "@/types/index";
 import { useQueryState, parseAsString } from "nuqs";
 import { useEffect, useState } from "react";
 import { ThreadItem } from "./thread-item";
